:truck: Move ChartConfigService files to shared folder
