+++ libs/designsystem/src/lib/components/charts/shared/chart-config-service/chart-config.service.ts
@@ -4,11 +4,11 @@ import { AnnotationType, AnnotationTypeRegistry } from 'chartjs-plugin-annotatio
 import { toDate } from 'date-fns';
 
 import { ChartDataLabelOptions, ChartLocale, ChartType, ChartTypeConfig } from '../../';
-import { deepCopy } from '../../../../../helpers/deep-copy';
+import { deepCopy } from '../../../../helpers/deep-copy';
 
-import { CHART_ANNOTATIONS_CONFIG } from './annotations.config';
-import { CHART_INTERACTION_FUNCTIONS_EXTENSIONS } from './interaction-functions-extensions.config';
-import { CHART_TYPES_CONFIG } from './type.config';
+import { CHART_ANNOTATIONS_CONFIG } from './configs/annotations.config';
+import { CHART_INTERACTION_FUNCTIONS_EXTENSIONS } from './configs/interaction-functions-extensions.config';
+import { CHART_TYPES_CONFIG } from './configs/type.config';
 
 @Injectable()
 export class ChartConfigService {
