+++ libs/designsystem/src/lib/components/charts/shared/chart-config-service/configs/annotations.config.ts
