+++ libs/designsystem/src/lib/components/charts/shared/chart-config-service/configs/global-defaults.config.ts
