+++ libs/designsystem/src/lib/components/charts/shared/chart-config-service/configs/interaction-functions-extensions.config.ts
