+++ libs/designsystem/src/lib/components/charts/shared/chart-config-service/configs/type.config.ts
