+++ libs/designsystem/src/lib/components/charts/shared/chart-config-service/index.ts
@@ -1,0 +1,1 @@
+export { ChartConfigService } from './chart-config.service';
