+++ libs/designsystem/src/lib/components/charts/shared/chart-js-service/chart-js.service.spec.ts
@@ -4,12 +4,12 @@ import { Chart, ChartOptions, ChartType as ChartJSType } from 'chart.js';
 import { AnnotationOptions } from 'chartjs-plugin-annotation';
 import { MockProvider } from 'ng-mocks';
 
+import { ChartConfigService } from '../';
 import { ChartDataset, ChartHighlightedElements, ChartType } from '../../';
 import { deepCopy } from '../../../../helpers/deep-copy';
+import { CHART_GLOBAL_DEFAULTS } from '../chart-config-service/configs/global-defaults.config';
 
 import { ChartJSService } from './chart-js.service';
-import { ChartConfigService } from './configs/chart-config.service';
-import { CHART_GLOBAL_DEFAULTS } from './configs/global-defaults.config';
 import { TEST_CHART_ANNOTATIONS_CONFIG, TEST_CHART_TYPES_CONFIG } from './test-utils';
 
 describe('ChartJSService', () => {
