+++ libs/designsystem/src/lib/components/charts/shared/chart-js-service/configured-chart-js.ts
@@ -17,9 +17,9 @@ import annotationPlugin from 'chartjs-plugin-annotation';
 import ChartDataLabels from 'chartjs-plugin-datalabels';
 
 import { mergeDeepAll } from '../../../../helpers/merge-deep';
+import { CHART_GLOBAL_DEFAULTS } from '../chart-config-service/configs/global-defaults.config';
 
 import MarkerPlugin from './chartjs-plugin-marker/chartjs-plugin-marker';
-import { CHART_GLOBAL_DEFAULTS } from './configs/global-defaults.config';
 
 const CHART_SCALES = [CategoryScale, LinearScale, TimeScale];
 const CHART_ELEMENTS = [BarElement, LineElement, PointElement];
