+++ libs/designsystem/src/lib/components/charts/shared/chart-js-service/index.ts
@@ -1,3 +1,2 @@
 export { ChartJSService } from './chart-js.service';
-export { ChartConfigService } from './configs/chart-config.service';
 export * from './test-utils';
