+++ libs/designsystem/src/lib/components/charts/shared/index.ts
@@ -1,3 +1,4 @@
 export * from './charts.types';
 export * from './base-chart';
 export * from './chart-js-service';
+export * from './chart-config-service';
